# symfony/src/Command/openapi.yaml
openapi: 3.1.0
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: 'http://swagger.io/terms/'
  contact:
    url: 'https://google.com'
    name: Bob Dylan
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.11
...
servers:
  - url: 'https://petstore3.swagger.io/api/v3'
...
paths:
  '/client/{clientId}/{param1}/{param2}/{param3}/{param4}/{param5}':
    get:
      description: desc
      summary: Your GET endpoint
      responses:
        '200':
          $ref: '#/components/responses/TestResp'
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    x-stoplight:
                      id: xyxfabzwnf6lc
      operationId: get-client
      requestBody:
        content:
          application/json:
            schema:
              type: integer
          application/xml:
            schema:
              type: object
              properties: {}
      x-priority: 1
      parameters:
        - schema:
            type: string
          in: header
          name: azef
          required: true
          description: desc
        - schema:
            type: number
          in: query
          name: agrez
          required: true
          description: desc
        - schema:
            type: integer
            default: 10
          in: cookie
          name: azgrzeg
          description: desc
        - schema:
            type: boolean
            default: true
          in: header
          name: gegzer
          description: desc
    parameters:
      - name: clientId
        in: path
        schema:
          type: string
        required: true
        description: desc
      - schema:
          format: format
          default: item
          enum:
            - item
            - item2
          example: item
        name: param1
        in: path
        required: true
        description: desc
      - schema:
          type: string
          format: format
          default: item
          enum:
            - item
            - item1
          example: item
          pattern: item
          minLength: 1
          maxLength: 10
        name: param2
        in: path
        required: true
        style: matrix
        description: desc
      - schema:
          type: number
          maximum: 2
          multipleOf: 1
        name: param3
        in: path
        required: true
        style: matrix
        description: desc
      - schema:
          type: integer
        name: param4
        in: path
        required: true
        description: desc
      - schema:
          type: boolean
        name: param5
        in: path
        required: true
        description: desc
    post:
      description: desc
      summary: ''
      operationId: post-client-clientId-param1-param2-param3-param4-param5-param6
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    x-stoplight:
                      id: x0sw75ce8wr5q
...
components:
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'https://petstore3.swagger.io/oauth/authorize'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header
  responses:
    TestResp:
      description: Example response
      content:
        application/json:
          schema:
            type:
              - object
            properties:
              test:
                type: string
                x-stoplight:
                  id: bd58jchxoa10a